ImportLoops: import called loops in .bedpe format to use in HicAggR

Imports bedpe file formats as GInteractions object usable to perform submatrix extraction with ExtractSubmatrix()

Usage

ImportLoops(
  file_bedpe = NULL,
  genomicConstraint = NULL,
  discard_trans = FALSE,
  chromSizes = NULL,
  binSize = NULL,
  minDist = NULL,
  maxDist = NULL,
  verbose = FALSE,
  cores = 1
)

Arguments

file_bedpe

bedpe file path (Default NULL)

genomicConstraint

: GRanges object of constraint regions. If NULL chromosomes in chromSizes are used as constraints (Default NULL).

discard_trans

: If TRUE discard loops where anchor and bait are in different genomic constraint elements, either different TADs or chromosomes, if genomicConstraint = NULL. (Default FALSE)

chromSizes

<data.frame>: A data.frame containing chromosomes names and lengths in base pairs (see example). (Default NULL)

binSize

: Bin size in bp - corresponds to matrix resolution.

minDist

: Minimal distance between anchors and baits. (Default NULL)

maxDist

: Maximal distance between anchors and baits. (Default NULL)

verbose

: Show the progression in console? (Default FALSE)

cores

: Number of cores to use. (Default 1)

Value

A GInteractions object.

Examples

bedpe_path <- system.file("extdata",
    "postprocessed_pixels_5000.bedpe",
    package = "HicAggR", mustWork = TRUE
)
loops <- ImportLoops(
    file_bedpe = bedpe_path,
    chromSizes = data.frame(c("1","2"),
        c(249250621,243199373)),
    binSize = 5000, minDist = 105000
)
#> Warning: found more than one best sequence renaming map compatible with seqname style "NCBI" for this object, using the first one